+++ tailwind.config.ts
@@ -1,5 +1,8 @@
+import typography from "@tailwindcss/typography";
+import daisyui from "daisyui";
+
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ["./app/**/*.{ts,tsx,jsx,js}"],
   theme: {
     fontFamily: {
@@ -11,7 +14,7 @@ module.exports = {
       },
     },
   },
-  plugins: [require("@tailwindcss/typography"), require("daisyui")],
+  plugins: [typography, daisyui],
   // daisyUI config (optional)
   daisyui: {
     styled: true,
